Remove user authentication and consolidate components
* **README.md**
  - Remove "User Authentication Setup" section
  - Update features list to remove "User authentication"

* **src/App.js**
  - Remove `auth` import, `useEffect` hook, and `onAuthStateChanged` function
  - Remove `Login` and `Signup` components from `Stack.Navigator`
  - Update `App` component to import combined library from `src/Library.js`

* **src/Library.js**
  - Combine all components and styles from `src/components` and `src/styles` into this file

* Delete `src/components` and `src/styles` files
