--- src/components/BookList.js
@@ -1,94 +1,0 @@
-import React, { useState } from 'react';
-import { View, Text, TextInput, Button, FlatList, TouchableOpacity, Alert } from 'react-native';
-
-const BookList = ({ navigation }) => {
-  const [books, setBooks] = useState([]);
-  const [searchTerm, setSearchTerm] = useState('');
-
-  const handleSearch = (text) => {
-    setSearchTerm(text);
-  };
-
-  const filteredBooks = books.filter((book) =>
-    book.title.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  const handleDelete = (id) => {
-    Alert.alert(
-      "Delete Book",
-      "Are you sure you want to delete this book?",
-      [
-        {
-          text: "Cancel",
-          style: "cancel"
-        },
-        {
-          text: "OK",
-          onPress: () => setBooks(books.filter((book) => book.id !== id))
-        }
-      ]
-    );
-  };
-
-  const renderItem = ({ item }) => (
-    <View style={styles.bookItem}>
-      <TouchableOpacity onPress={() => navigation.navigate('BookDetails', { id: item.id })}>
-        <Text style={styles.bookTitle}>{item.title}</Text>
-      </TouchableOpacity>
-      <Button title="Delete" onPress={() => handleDelete(item.id)} />
-    </View>
-  );
-
-  return (
-    <View style={styles.container}>
-      <Text style={styles.header}>My Library</Text>
-      <TextInput
-        style={styles.searchInput}
-        placeholder="Search books..."
-        value={searchTerm}
-        onChangeText={handleSearch}
-      />
-      <FlatList
-        data={filteredBooks}
-        renderItem={renderItem}
-        keyExtractor={(item) => item.id.toString()}
-      />
-      <Button title="Add Book" onPress={() => navigation.navigate('AddBook')} />
-    </View>
-  );
-};
-
-const styles = {
-  container: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: '#f9f9f9',
-  },
-  header: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-  },
-  searchInput: {
-    height: 40,
-    borderColor: '#ccc',
-    borderWidth: 1,
-    borderRadius: 4,
-    paddingHorizontal: 10,
-    marginBottom: 20,
-  },
-  bookItem: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 10,
-    borderBottomWidth: 1,
-    borderBottomColor: '#ccc',
-  },
-  bookTitle: {
-    fontSize: 18,
-    color: '#4CAF50',
-  },
-};
-
-export default BookList;
